+++ src/main/java/com/springcooler/sgma/recruitmentboard/command/application/dto/RecruitmentBoardCommandDTO.java
@@ -1,0 +1,25 @@
+package com.springcooler.sgma.recruitmentboard.command.application.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.sql.Timestamp;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class RecruitmentBoardCommandDTO {
+    private Long recruitmentBoardId;
+    private String title;
+    private Timestamp createdAt;
+    private Timestamp updatedAt;
+    private Timestamp recruitmentStartTime;
+    private Timestamp recruitmentEndTime;
+    private String activeStatus;
+    private int likes;
+    private long groupId;
+    private int studyGroupCategoryId;
+}
