+++ src/main/java/com/springcooler/sgma/recruitmentboard/command/domain/aggregate/BoardActiveStatus.java
@@ -1,0 +1,5 @@
+package com.springcooler.sgma.recruitmentboard.command.domain.aggregate;
+
+public enum BoardActiveStatus {
+    ACTIVE, INACTIVE, DELETED
+}
