+++ src/main/java/com/springcooler/sgma/recruitmentboard/command/domain/service/StudyGroupApplicantDomainService.java
@@ -1,0 +1,4 @@
+package com.springcooler.sgma.recruitmentboard.command.domain.service;
+
+public class StudyGroupApplicantDomainService {
+}
