+++ src/main/java/com/springcooler/sgma/recruitmentboard/command/infrastructure/service/StudyGroupApplicantService.java
@@ -1,0 +1,4 @@
+package com.springcooler.sgma.recruitmentboard.command.infrastructure.service;
+
+public class StudyGroupApplicantService {
+}
