--- src/main/java/com/springcooler/sgma/studygroupapplicant/command/domain/aggregate/BoardActiveStatus.java
@@ -1,5 +1,0 @@
-package com.springcooler.sgma.studygroupapplicant.command.domain.aggregate;
-
-public enum BoardActiveStatus {
-    ACTIVE, INACTIVE, DELETED
-}
