--- src/main/java/com/springcooler/sgma/studygroupapplicant/command/domain/service/StudyGroupApplicantDomainService.java
@@ -1,4 +1,0 @@
-package com.springcooler.sgma.studygroupapplicant.command.domain.service;
-
-public class StudyGroupApplicantDomainService {
-}
