--- src/main/java/com/springcooler/sgma/studygroupapplicant/command/infrastructure/service/StudyGroupApplicantService.java
@@ -1,4 +1,0 @@
-package com.springcooler.sgma.studygroupapplicant.command.infrastructure.service;
-
-public class StudyGroupApplicantService {
-}
