--- src/main/java/com/springcooler/sgma/studygroupapplicant/query/repository/StudyGroupApplicantMapper.java
@@ -1,14 +1,0 @@
-package com.springcooler.sgma.studygroupapplicant.query.repository;
-
-import com.springcooler.sgma.studygroupapplicant.query.dto.StudyGroupApplicantDTO;
-
-
-import java.util.List;
-
-
-public interface StudyGroupApplicantMapper {
-    List<StudyGroupApplicantDTO> studyGroupRecruitment();
-
-    StudyGroupApplicantDTO selectStudyGroupApplicantDTO(Long recruitmentBoardId);
-
-}
